+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDProfileId.cs
@@ -148,6 +148,8 @@ namespace UnityEngine.Rendering.HighDefinition
         XRMirrorView,
         XRCustomMirrorView,
         XRDepthCopy,
+        BuildMaginals,
+        BuildMaginalInternal,
 
         // Low res transparency
         DownsampleDepth,