+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/RenderPipelineResources.cs
@@ -159,12 +159,26 @@ namespace UnityEngine.Rendering.HighDefinition
             public ComputeShader encodeBC6HCS;
             [Reload("Runtime/Core/CoreResources/CubeToPano.shader")]
             public Shader cubeToPanoPS;
+            [Reload("Runtime/Core/CoreResources/CubeToHemiPano.shader")]
+            public Shader cubeToHemiPanoPS;
             [Reload("Runtime/Core/CoreResources/BlitCubeTextureFace.shader")]
             public Shader blitCubeTextureFacePS;
             [Reload("Runtime/Material/LTCAreaLight/FilterAreaLightCookies.shader")]
             public Shader filterAreaLightCookiesPS;
             [Reload("Runtime/Core/CoreResources/ClearUIntTextureArray.compute")]
             public ComputeShader clearUIntTextureCS;
+            [Reload("Runtime/Core/CoreResources/GPUScan.compute")]
+            public ComputeShader gpuScanCS;
+            [Reload("Runtime/Core/CoreResources/Arithmetics.compute")]
+            public ComputeShader gpuArithmeticsCS;
+            [Reload("Runtime/Core/CoreResources/Rescale01.compute")]
+            public ComputeShader rescale01CS;
+            [Reload("Runtime/Core/CoreResources/ImportanceSample2D.compute")]
+            public ComputeShader importanceSample2DCS;
+            [Reload("Runtime/Core/CoreResources/ImportanceSamplingFromSamples.compute")]
+            public ComputeShader ImportanceSamplingFromSamplesCS;
+            [Reload("Runtime/Core/CoreResources/InverseCDF1D.compute")]
+            public ComputeShader inverseCDF1DCS;
 
             // XR
             [Reload("Runtime/ShaderLibrary/XRMirrorView.shader")]